Reimplement iset-search with one-pass semantics, full bitmap support.
While bitmap support for iset-search was supposed to be
added in bf574b56, a number of cases contained errors.
This corrects those cases and simplifies the implementation
overall.

It is also an error to pass the `update' continuation a key
which is not eqv? to the key that was searched.  This follows
the intended semantics of *-search procedures, but may have to
be adjusted depending on the final contents of the SRFI doc.